LP-126 Refactor directiories for notifications
