+++ mobileapp/src/components/screens/notification/NotificationListScreen/NotificationListEntry.tsx
@@ -3,7 +3,7 @@ import { Box, Divider, HStack, Pressable, Text } from 'native-base'
 import { NotificationDetails } from 'webService/interface/notification'
 import { StackNavigationProp } from '@react-navigation/stack'
 import { SignInRouterProps } from 'components/routers/RouterPropTypes'
-import { NotificationIcon } from 'components/screens/notification/NotificationIcon'
+import { NotificationIcon } from 'components/screens/notification/NotificationListScreen/NotificationIcon'
 import R from 'res/R'
 
 type Props = {
